parseInt > parseFloat

+++ models/models.js
@@ -59,7 +59,7 @@ module.exports.Homepage_image = model('Homepage_image', Schema({
 
 module.exports.Product = model('Product', Schema({
     name: String,
-    price: { type: Number, set: n => parseInt(n) * 100 },
+    price: { type: Number, set: n => parseFloat(n) * 100 },
     image: String,
     info: String,
     stock_qty: { type: Number, min: [0, "No negative values allowed for stock quantity"] }
